populate item prices using integers and populate cart fragment correctly

+++ web/app/syzygy/order_items/model.py
@@ -28,7 +28,7 @@ class OrderItem(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     quantity = db.Column(db.Integer)
-    price = db.Column(db.Numeric(10, 2))
+    price = db.Column(db.Integer)
 
     itemid = db.Column(db.Integer, db.ForeignKey("items.id"))
     item = db.relationship("Item")
